Support Vim 8 Packages
First major update in long time. This switches infect from working with
Pathogen to the built-in Packages.

This introduces two new commands that you can add to your vimrc file:

- `package`
- `plugin`

`bundle` command is not deprecated.

A package should point to a repo that has the file structure needed for
the vim packages, including nesting files under `start/` and/or `opt/`.
Normal plugins can still be installed directly, and infect will take
care of creating the proper directories for you. By default it will
create a `default` package and add everything to the auto-loaded
`start/` directory.

In other words, it should work just like pathogen did.

New addition is the `build` directive to a command. Now infect can run a
command in the directory of the plugin to perform any build steps
needed. E.g. to run `make` in vimproc I would add this line:

    "=plugin Shougo/vimproc.vim build: make

Removal of multi-threaded runner. I'd like to add this back in eventually, but
I removed it during the refactor to simplify the code. Ideally I would
use something like ncurses to have a reasonable output while All The
Things™ were run.™ were run.

Note: Pathogen should still work, but would require pointing the install
directory to use the 'pack' directory. The only reason I think you'd
want to do that would be if you had a version of vim earlier than 8 and
still wanted to use infect. The updated lines in your vimrc might look
like this:

    execute pathogen#infect('pack/default/auto/{}')