Prompt for uninstall properly
When globbing for `*/*/*` under the pack directory we could potentially
prompt for every file that is part of a proper vim package, which made
the infect cleanup process noisy to the point of uselessness.

Now will search two directorys:

* `.vim/pack/plugins/*/*`
* `.vim/pack/*`

But not `.vim/pack/plugins`, since that is our default "legacy plugin"
folder.

There should only ever be "auto" and "opt" at the root of
`.vim/pack/plugins/*` so I just globbed for everything there.